pyproject.toml
[tool.poetry]
name = "instagram-scraper"
version = "0.1.0"
description = "Custom IG scraper"
authors = ["Your Name <email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
httpx = "^0.25.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
fake-useragent = "^1.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-httpx = "^0.24.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]